web/DownloadButton: extract api interaction logic into a lib
download button state is now stored, well, in a state
