+++ web/src/lib/state/omnibox.ts
@@ -1,3 +1,5 @@
 import { writable } from "svelte/store";
+import type { CobaltDownloadButtonState } from "$lib/types/omnibox";
 
 export const link = writable("");
+export const downloadButtonState = writable<CobaltDownloadButtonState>("idle");
