+++ web/src/lib/types/omnibox.ts
@@ -1,0 +1,1 @@
+export type CobaltDownloadButtonState = "idle" | "think" | "check" | "done" | "error";
